Post -> PostDto 변경에 따른 코드 수정

+++ src/main/java/com/github/supercodingfinalprojectbackend/controller/PostController.java
@@ -26,7 +26,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}")
-    public ResponseEntity<ApiResponse<PostDto>> getPost(@PathVariable Integer postId){
+    public ResponseEntity<ApiResponse<PostDto>> getPost(@PathVariable Long postId){
         return postService.getPost(postId);
     }
 }
